Editing of classes per student
This is pretty much the most minimal implementation I could think
of. Since we mostly need this to be able to reference classes "later"
they are now implemented as string properties of student aggregates.

This is a little complicated by the fact that classes as a grouping of
students can also be seen as subsets of departments which are subsets of
schools. Which is why you can only set a class name for students that
already have a department and class names get erased when you move
students to other departments.