State transitions for sections are now strict
An event does change the state of the section only if the old state is
the one specified in the event-handler. This ensures that states
transitions always happens in the same way.

This required a fix in the tests because now you can't go anymore from
nil to finished, you need to first assign a question (and go in-progress)