Preheat the aggregate cache with course material
This change ensures that any course material that's been published
will be in the aggregate cache before the web app is made available.

This is important because loading the course material aggregate takes a
fairly long time and quite a lot of memory. If we don't pre cache this
it can bring down the application when we get a bunch of commands that
interact with the course material immediately after startup.