Unify & standardize ring session handling
Use a standard, shared ring session for all applications. The session is
stored in redis in production and in-memory in development.

This replaces both the custom session store for credentials and the
separate login "temporary" session store which was not shared between
login servers.

This change should fix the login issues @
https://app.asana.com/0/9685197724612/16989567729100