Allow teachers to use learning environment.
Surprisingly, as far as I can tell this tiny commit works and does not
seem to cause any real difficulty. Mostly because no events in the
learning environment apply directly to student aggregates.

Note that this does change the startup arguments for learning and
teaching environments meaning we have to be careful about deploying.