Remove unsupported syntax
Keywords like ::some-alias/foo work in the clojure reader, but only if
`some-alias` is an alias of a namespace in the current package. That
means that source code that uses keywords like that cannot be read
without compiling. So static analysis breaks on that code.

As far as I can tell these aliased keywords were never intended to work
at all so let's get rid of them and gain some interoperability at the
cost of verbosity.